requirements.txt
# Core
aiohttp==3.9.1
aiosqlite==0.19.0
sqlalchemy==2.0.23
pydantic>=2.5.2
python-dotenv==1.0.0
python-telegram-bot==20.6
pydantic-settings

# Data & Utils
numpy==1.26.3
pandas==2.1.4
pyyaml==6.0.1
loguru==0.7.2
colorama==0.4.6